+++ benchmark/suites/events.js
@@ -21,6 +21,18 @@ let bench = benchmark.createSuite("Emit event");
 	// Create broker
 	let broker = new ServiceBroker();
 
+	broker.on("event", p => p);
+
+	bench.add("Emit simple event to 1 subscribers", () => {
+		return broker.emit("event", ["param1", { a: 1, b: "Teszt"}, 500]);
+	});
+
+})();
+
+(function() {
+	// Create broker
+	let broker = new ServiceBroker();
+
 	for(let i = 0; i < 20; i++) 
 		broker.on("event", p => p);
 
@@ -85,17 +97,19 @@ Platform info:
    Intel(R) Core(TM) i7-4770K CPU @ 3.50GHz × 8
 
 Suite: Emit event
-√ Emit event without subscribers x 3,654,151 ops/sec ±1.53% (94 runs sampled)
-√ Emit simple event to 20 subscribers x 1,043,375 ops/sec ±0.99% (92 runs sampled)
-√ Emit wildcard event to 20 subscribers x 945,761 ops/sec ±0.64% (93 runs sampled)
-√ Emit multi-wildcard event to 20 subscribers without params x 855,044 ops/sec ±0.69% (95 runs sampled)
-√ Emit multi-wildcard event to 20 subscribers with params x 853,537 ops/sec ±0.65% (95 runs sampled)
+√ Emit event without subscribers                                     2,161,567 rps
+√ Emit simple event to 1 subscribers                                 1,785,380 rps
+√ Emit simple event to 20 subscribers                                1,076,511 rps
+√ Emit wildcard event to 20 subscribers                                977,945 rps
+√ Emit multi-wildcard event to 20 subscribers without params           894,379 rps
+√ Emit multi-wildcard event to 20 subscribers with params              880,544 rps
 
-   Emit event without subscribers                                 0.00%   (3,654,151 ops/sec)
-   Emit simple event to 20 subscribers                          -71.45%   (1,043,375 ops/sec)
-   Emit wildcard event to 20 subscribers                        -74.12%    (945,761 ops/sec)
-   Emit multi-wildcard event to 20 subscribers without params   -76.60%    (855,044 ops/sec)
-   Emit multi-wildcard event to 20 subscribers with params      -76.64%    (853,537 ops/sec)
+   Emit event without subscribers (#)                                    0%      (2,161,567 rps)   (avg: 462ns)
+   Emit simple event to 1 subscribers                                -17.4%      (1,785,380 rps)   (avg: 560ns)
+   Emit simple event to 20 subscribers                               -50.2%      (1,076,511 rps)   (avg: 928ns)
+   Emit wildcard event to 20 subscribers                            -54.76%        (977,945 rps)   (avg: 1μs)
+   Emit multi-wildcard event to 20 subscribers without params       -58.62%        (894,379 rps)   (avg: 1μs)
+   Emit multi-wildcard event to 20 subscribers with params          -59.26%        (880,544 rps)   (avg: 1μs)
 -----------------------------------------------------------------------
 
 */